feat: publish and redirect webhook and document actions
