+++ src/sanity/actions.tsx
@@ -2,10 +2,8 @@
 /* eslint-disable no-negated-condition */
 import { useToast } from "@sanity/ui";
 import React from "react";
-import { useEffect, useState } from "react";
 import { BsSend } from "react-icons/bs";
 import { DocumentActionComponent, useDocumentOperation } from "sanity";
-import { DocumentActionProps, useClient, useValidationStatus } from "sanity";
 
 import { ValueType } from "@/types/types";
 
@@ -128,64 +126,3 @@ export const SendToIterable: DocumentActionComponent = ({ id, type }) => {
     disabled: isDocumentPublished,
   };
 };
-
-/**
- * publish and redirect action
- */
-export default function PublishAndRedirect(props: DocumentActionProps) {
-  const client = useClient();
-  const toast = useToast();
-  const { publish } = useDocumentOperation(props.id, props.type);
-  const validation = useValidationStatus(props.id, props.type);
-  // Check if the document is valid before allowing the user to publish it.
-  const isValid =
-    !validation.isValidating &&
-    validation.validation.filter((v) => v.level === "error").length === 0;
-
-  const [isPublishing, setIsPublishing] = useState(false);
-  const draftPath = resolveProductionPath(props.draft);
-  const publishedPath = resolveProductionPath(props.published);
-
-  useEffect(() => {
-    // if the isPublishing state was set to true and the draft has changed
-    // to become `null` the document has been published
-    if (isPublishing && !props.draft) {
-      setIsPublishing(false);
-    }
-  }, [props.draft]);
-
-  const handlePublish = async () => {
-    // This will update the button text
-    setIsPublishing(true);
-    // Create the redirect
-    await client.create({
-      _type: "redirect",
-      source: publishedPath,
-      destination: draftPath,
-      permanent: true,
-    });
-
-    // Notify the user the redirect was created.
-    toast.push({
-      closable: true,
-      duration: 5000,
-      status: "success",
-      title: `Redirect created`,
-      description: `Redirect from ${publishedPath} to ${draftPath}`,
-    });
-    // Perform the publish
-    publish.execute();
-    // Signal that the action is completed
-    props.onComplete();
-  };
-  if (publishedPath && draftPath && draftPath !== publishedPath) {
-    return {
-      disabled: !!publish.disabled || !isValid,
-      tone: "positive",
-      label: isPublishing ? "Publishing…" : "Publish & Redirect",
-      onHandle: handlePublish,
-    } as const;
-  }
-
-  return null;
-}
